requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0

# AI / LLM
langgraph>=0.2.60
langchain-anthropic>=0.3.4
langchain-core>=0.3.31

# Vietnam Stock Data
vnstock>=3.4.0
pytz>=2024.1

# Database
supabase>=2.11.0

# Scheduler
apscheduler>=3.10.4

# Telegram
python-telegram-bot>=21.9

# Data Processing
pandas>=2.2.0
numpy>=2.1.0
ta>=0.11.0

# Backtesting
backtrader>=1.9.78
matplotlib>=3.8.0

# DeepSeek R1 (OpenAI-compatible API)
langchain-openai>=0.3.0
